Refactored the namespace.
Added utilitary methods.

Created the IGrammar base class to be used as the base for all the
syntax parsers.

Created the IGrammarRule base class to be used as the base for all
the grammar rule parsers.

Implemented a recursive descent LL(*) parser "LLStar".

Implemented the complete YAML Language File parser "Language".

Changed the tokenizer to create an Abstract Syntax Tree automatically
from the provided grammar rules.

Added a method to the compiled C99 code to translate a symbol id to
its name.