Created the LLStar parser generator Unit Test.
Added an extra helper function to the Code class to invert the siblings of
an AST tree.

Created a new method on the Language class to get the Id of a Symbol.
Organized and commented the Language class code.

Bug correction on the LLStar class, it was returning the AST tree with
the same level nodes in the inverse order.